fix room
