+++ src/main/java/com/hospital/appointmentservice/patient/entity/MedicalVisit.java
@@ -48,6 +48,11 @@ public class MedicalVisit {
 
     @Column(name = "created_at")
     private LocalDateTime createdAt;
+
+    //Update
+    @Column(name = "status")
+    private String status;
+
 }
