+++ src/main/java/com/hospital/appointmentservice/receptionist/controller/ReceptionistController.java
@@ -1,18 +1,15 @@
 package com.hospital.appointmentservice.receptionist.controller;
 
-import com.hospital.appointmentservice.receptionist.dto.AppointmentResponse;
-import com.hospital.appointmentservice.receptionist.dto.ReceptionistDto;
-import com.hospital.appointmentservice.receptionist.service.ReceptionistService;
+
+import com.hospital.appointmentservice.receptionist.service.ReceptionistPatientService;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/receptionist")
@@ -21,7 +18,7 @@ import java.util.List;
 // Annotation make field private and final
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class ReceptionistController {
-    ReceptionistService receptionistService;
+    ReceptionistPatientService receptionistService;
 
     @GetMapping("")
     public ResponseEntity<?> getAllReceptionists() {
