--- src/main/java/com/hospital/appointmentservice/receptionist/entity/MedicalRecord.java
@@ -1,35 +1,0 @@
-package com.hospital.appointmentservice.receptionist.entity;
-
-import com.hospital.appointmentservice.admin.model.Appointment;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-@Entity
-@Table(name = "MedicalRecords_New")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MedicalRecord {
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(name = "id", nullable = false)
-    private UUID id;
-
-    @Column(columnDefinition = "NVARCHAR(MAX)")
-    private String diagnosis;
-
-    @Column(columnDefinition = "NVARCHAR(MAX)")
-    private String notes;
-
-    private LocalDateTime createdAt = LocalDateTime.now();
-
-    @OneToOne
-    @JoinColumn(name = "appointmentId", nullable = false, unique = true)
-    private Appointment appointment;
-}
-
-
