+++ src/main/java/com/hospital/appointmentservice/receptionist/service/RoomService.java
@@ -16,7 +16,7 @@ import org.springframework.data.domain.Pageable;
 
 @Service
 public class RoomService {
-    // test push lên GitHub
+
     @Autowired
     private RoomReceptionRepository roomRepository;
